All urllib replaced with requests, most location manager functions moved to location_manager.py, all HTTP requests within try/except blocks.
